Remove (most) Python 2 support from CI scripts
Summary:
Python 2 has reached end-of-life and is no longer supported by PyTorch.
Now we can clean up a lot of cruft that we put in place to support it.
These changes were all done manually, and I skipped anything that seemed
like it would take more than a few seconds, so I think it makes sense to
review it manually as well.

Test Plan:
CI

ghstack-source-id: 7c25535e521afbf6aaf88a8d38177411f03bf26d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35616