qactivation_benchmarks: small bug fix
Summary:

Changes relu and relu6 to point to the functional implementations here.
The previous behavior tested the time to create the module, but didn't actually run the
function (I noticed this when adding the new input sizes and seeing
the measured time not change).

Test Plan:

run the benchmark, the time now changes as expected with input size for
these.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: c920fb11ee680d42c17a0c1f677f1c71a44e1e91
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35731