Avoid one unnecessary memory allocation and copy in XNNPACK integration.
Currently we call input.contiguous() on the input tensor resulting in an
unecessary allocation and copy in cases where the input is not contiguous
with regards to the requested memory format.  The reason is that in such
scenarios, this call re-allocates and copies the input tensor into
contiguous storage, only for this newly allocated tensor to be used as
the source of another copy to the final destination.  Instead, if we copy
into the destination directly in such circumstances, we will save an
allocation and a copy.

ghstack-source-id: 2a6a4cf90463c7f55d15bc226ff9677f0535ab8c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35350