Enable stateless XNNPACK convolutions.
The optimal solution to use XNNPACK is to separate operator creation
from execution - also called prepacking the weights.  If we have done
our job properly, JIT must have caught and replaced nn.Conv2ds on mobile
with the prepacked versions.  Still, if we somehow end up in
_convolution for whatever reason, it is still more efficient to go
through XNNPACK for NHWC tensors, compared to the alternative of
converting NHWC to NCHW and going through NNPACK.

ghstack-source-id: 180a126f46110bca6643c92bdda2c6a48dacfa22
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35790